refactoring playerInput
